feat(desktop): add troubleshooting settings section
